añadida la linea protected  = true

+++ database/migrations/2026_01_28_000002_create_users_idiomas_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->unsignedBigInteger('idioma_id');
+            $table->unsignedInteger('idioma_id');
             $table->foreign('idioma_id')->references('id')->on('idiomas')->onDelete('cascade');
             $table->timestamps();
         });
